+++ Src/PerformanceCollector/Shared/Implementation/PerformanceCounterData.cs
@@ -17,10 +17,26 @@
             this.UsesInstanceNamePlaceholder = usesInstanceNamePlaceholder;
             this.IsCustomCounter = isCustomCounter;
             this.IsInBadState = isInBadState;
-            this.CategoryName = categoryName;
-            this.CounterName = counterName;
-            this.InstanceName = instanceName;
+            this.PerformanceCounter = new PerformanceCounterStructure(categoryName, counterName, instanceName);
         }
+
+        public PerformanceCounterData(
+            string originalString,
+            string reportAs,
+            bool usesInstanceNamePlaceholder,
+            bool isCustomCounter,
+            bool isInBadState,
+            PerformanceCounterStructure counter)
+        {
+            this.OriginalString = originalString;
+            this.ReportAs = reportAs;
+            this.UsesInstanceNamePlaceholder = usesInstanceNamePlaceholder;
+            this.IsCustomCounter = isCustomCounter;
+            this.IsInBadState = isInBadState;
+            this.PerformanceCounter = counter;
+        }
+
+        public PerformanceCounterStructure PerformanceCounter { get; private set; }
 
         public string OriginalString { get; private set; }
 
@@ -31,11 +47,5 @@
         public bool IsCustomCounter { get; private set; }
 
         public bool IsInBadState { get; private set; }
-
-        public string CategoryName { get; private set; }
-
-        public string CounterName { get; private set; }
-
-        public string InstanceName { get; private set; }
     }
 }